+++ driver/src/main/java/oracle/nosql/driver/kv/StoreAccessTokenProvider.java
@@ -471,10 +471,12 @@ public class StoreAccessTokenProvider implements AuthorizationProvider {
         try {
             final HttpHeaders headers = new DefaultHttpHeaders();
             headers.set(AUTHORIZATION, authHeader);
-            client = new HttpClient(
-                loginHost, loginPort, 0, 0, 0,
-                (isSecure && !disableSSLHook) ? sslContext : null,
-                serviceName, logger);
+            client = HttpClient.createMinimalClient
+                (loginHost,
+                 loginPort,
+                 (isSecure && !disableSSLHook) ? sslContext : null,
+                 serviceName,
+                 logger);
             return HttpRequestUtil.doGetRequest(
                 client,
                 NoSQLHandleConfig.createURL(endpoint, basePath + serviceName)